TIMS-257 - changed free space calculation to use Java File class

+++ c2mon-shared/c2mon-shared-persistence-manager/src/java/cern/c2mon/pmanager/fallback/manager/FallbackFileManager.java
@@ -166,7 +166,7 @@ public class FallbackFileManager {
                                 + fFileController.getNumberOfLines()
                                 + " and the current value for nextcheck is "
                                 + nextCheckFallbackData);
-            int freeSpace = SystemResourcesParameters.getFreeSpace();
+            long freeSpace = SystemResourcesParameters.getFreeSpace();
             if (freeSpace < minimumFreeSpace
                     && freeSpace != FallbackProperties.CMD_FREE_SPACE_ERROR) {
                 throw new SystemDiskSpaceException("Reached the limit of free space of "
